+++ lib/elixir/lib/kernel/cli.ex
@@ -224,7 +224,7 @@ defmodule Kernel.CLI do
   end
 
   defp process_command({:eval, expr}, _config) when is_binary(expr) do
-    Code.eval(expr, [])
+    Code.eval_string(expr, [])
     :ok
   end