+++ lib/elixir/test/elixir/system_test.exs
@@ -39,7 +39,7 @@ defmodule SystemTest do
 
   test :argv do
     list = elixir('-e "IO.inspect System.argv" -- -o opt arg1 arg2 --long-opt 10')
-    { args, _ } = Code.eval list, []
+    { args, _ } = Code.eval_string list, []
     assert args == ["-o", "opt", "arg1", "arg2", "--long-opt", "10"]
   end